Testing moving several core functions and classes to separate file
